requirements.txt
asgiref==3.2.10
boto3==1.21.27
botocore==1.24.27
Django==3.1
django-filter==21.1
django-storages==1.12.3
docutils==0.16
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.0.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.5.2
psycopg2==2.9.2
dj-database-url==0.5.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==6.0.0
xhtml2pdf==0.2.6
imageio==2.8.0
imageio-ffmpeg==0.4.1